requirements.txt
dj-database-url==0.5.0
Django==2.0.8
gunicorn==19.9.0
Pillow==5.2.0
pytz==2018.5
whitenoise==4.1
psycopg2==2.7.5
django-crispy-forms==1.7.2
